fix(frontend): rename the folder components to pages
